This class manages a common functionalities for both multiparty and two party communications.
The class is called from the communication setup classes and does the creation of the channels and the connect step during the communication setup.

Although it is declared "public" it is not for public usage and we recommend not to use it directly. In order to setup a communication use one of the communication setup classes.